requirements.txt
arrow==0.15.4
astroid==1.6.6
attrs==19.3.0
aws-lambda-builders==0.6.0
aws-sam-cli==0.38.0
aws-sam-translator==1.19.0
backcall==0.1.0
binaryornot==0.4.4
boto3==1.10.41
botocore==1.13.41
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chevron==0.13.1
Click==7.0
cookiecutter==1.6.0
dateparser==0.7.2
decorator==4.4.1
Django==2.0.2
docker==4.1.0
docutils==0.15.2
Flask==1.0.4
future==0.18.2
git-remote-codecommit==1.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
ikp3db==1.4.1
importlib-metadata==1.3.0
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
jsonschema==3.2.0
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.1.1
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
regex==2019.12.19
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
serverlessrepo==0.1.9
six==1.14.0
stevedore==1.32.0
tomlkit==0.5.8
traitlets==4.3.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
websocket-client==0.56.0
Werkzeug==0.16.0
whichcraft==0.6.1
wrapt==1.12.0
zipp==0.6.0